feat: serve static files without vite
